bt-108: * refactoring imports and folders
