+++ mobile/src/bundles/talent/enums/enums.ts
@@ -1,0 +1,1 @@
+export { BadgeType } from './ui/ui';
