+++ mobile/src/bundles/talent/enums/ui/badge-type.enum.ts
