+++ mobile/src/bundles/talent/enums/ui/ui.ts
