+++ mobile/src/bundles/users/enums/enums.ts
@@ -1,2 +1,1 @@
-export { BadgeType } from './ui/ui';
 export { UsersApiPath } from 'shared/build/index';
